% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.greengrass_operations.R
\name{associate_service_role_to_account}
\alias{associate_service_role_to_account}
\title{Associates a role with your account}
\usage{
associate_service_role_to_account(RoleArn = NULL)
}
\arguments{
\item{RoleArn}{The ARN of the service role you wish to associate with your account.}
}
\description{
Associates a role with your account. AWS Greengrass will use the role to access your Lambda functions and AWS IoT resources. This is necessary for deployments to succeed. The role must have at least minimum permissions in the policy ''AWSGreengrassResourceAccessRolePolicy''.
}
\section{Accepted Parameters}{
\preformatted{associate_service_role_to_account(
  RoleArn = "string"
)
}
}

