% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.greengrass_operations.R
\name{create_deployment}
\alias{create_deployment}
\title{Creates a deployment}
\usage{
create_deployment(GroupId, AmznClientToken = NULL, DeploymentId = NULL,
  DeploymentType = NULL, GroupVersionId = NULL)
}
\arguments{
\item{GroupId}{The ID of the AWS Greengrass group.}

\item{AmznClientToken}{A client token used to correlate requests and responses.}

\item{DeploymentId}{The ID of the deployment if you wish to redeploy a previous deployment.}

\item{DeploymentType}{The type of deployment. When used in ''CreateDeployment'', only ''NewDeployment'' and ''Redeployment'' are valid.}

\item{GroupVersionId}{The ID of the group version to be deployed.}
}
\description{
Creates a deployment.
}
\examples{

}
