% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.greengrass_operations.R
\name{create_device_definition_version}
\alias{create_device_definition_version}
\title{Creates a version of a device definition that has already been defined}
\usage{
create_device_definition_version(
  AmznClientToken = "string",
  DeviceDefinitionId = "string",
  Devices = list(
    list(
      CertificateArn = "string",
      Id = "string",
      SyncShadow = TRUE|FALSE,
      ThingArn = "string"
    )
  )
)
}
\arguments{
\item{AmznClientToken}{A client token used to correlate requests and responses.}

\item{DeviceDefinitionId}{[required] The ID of the device definition.}

\item{Devices}{A list of devices in the definition version.}
}
\description{
Creates a version of a device definition that has already been defined.
}
\examples{

}
