% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.greengrass_operations.R
\name{create_function_definition_version}
\alias{create_function_definition_version}
\title{Creates a version of a Lambda function definition that has already been defined}
\usage{
create_function_definition_version(AmznClientToken = NULL,
  FunctionDefinitionId, Functions = NULL)
}
\arguments{
\item{AmznClientToken}{A client token used to correlate requests and responses.}

\item{FunctionDefinitionId}{[required] The ID of the Lambda function definition.}

\item{Functions}{A list of Lambda functions in this function definition version.}
}
\description{
Creates a version of a Lambda function definition that has already been defined.
}
\section{Accepted Parameters}{
\preformatted{create_function_definition_version(
  AmznClientToken = "string",
  FunctionDefinitionId = "string",
  Functions = list(
    list(
      FunctionArn = "string",
      FunctionConfiguration = list(
        EncodingType = "binary"|"json",
        Environment = list(
          AccessSysfs = TRUE|FALSE,
          ResourceAccessPolicies = list(
            list(
              Permission = "ro"|"rw",
              ResourceId = "string"
            )
          ),
          Variables = list(
            "string"
          )
        ),
        ExecArgs = "string",
        Executable = "string",
        MemorySize = 123,
        Pinned = TRUE|FALSE,
        Timeout = 123
      ),
      Id = "string"
    )
  )
)
}
}

