% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.greengrass_operations.R
\name{create_group}
\alias{create_group}
\title{Creates a group}
\usage{
create_group(AmznClientToken = NULL, InitialVersion = NULL,
  Name = NULL)
}
\arguments{
\item{AmznClientToken}{A client token used to correlate requests and responses.}

\item{InitialVersion}{Information about the initial version of the group.}

\item{Name}{The name of the group.}
}
\description{
Creates a group. You may provide the initial version of the group or use ''CreateGroupVersion'' at a later time.
}
\section{Accepted Parameters}{
\preformatted{create_group(
  AmznClientToken = "string",
  InitialVersion = list(
    CoreDefinitionVersionArn = "string",
    DeviceDefinitionVersionArn = "string",
    FunctionDefinitionVersionArn = "string",
    LoggerDefinitionVersionArn = "string",
    ResourceDefinitionVersionArn = "string",
    SubscriptionDefinitionVersionArn = "string"
  ),
  Name = "string"
)
}
}

