% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.greengrass_operations.R
\name{create_group_certificate_authority}
\alias{create_group_certificate_authority}
\title{Creates a CA for the group}
\usage{
create_group_certificate_authority(GroupId, AmznClientToken = NULL)
}
\arguments{
\item{GroupId}{The ID of the AWS Greengrass group.}

\item{AmznClientToken}{A client token used to correlate requests and responses.}
}
\description{
Creates a CA for the group. If a CA already exists, it will rotate the existing CA.
}
\examples{

}
