% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.greengrass_operations.R
\name{create_group_version}
\alias{create_group_version}
\title{Creates a version of a group which has already been defined}
\usage{
create_group_version(GroupId, AmznClientToken = NULL,
  CoreDefinitionVersionArn = NULL, DeviceDefinitionVersionArn = NULL,
  FunctionDefinitionVersionArn = NULL, LoggerDefinitionVersionArn = NULL,
  ResourceDefinitionVersionArn = NULL,
  SubscriptionDefinitionVersionArn = NULL)
}
\arguments{
\item{GroupId}{The ID of the AWS Greengrass group.}

\item{AmznClientToken}{A client token used to correlate requests and responses.}

\item{CoreDefinitionVersionArn}{The ARN of the core definition version for this group.}

\item{DeviceDefinitionVersionArn}{The ARN of the device definition version for this group.}

\item{FunctionDefinitionVersionArn}{The ARN of the function definition version for this group.}

\item{LoggerDefinitionVersionArn}{The ARN of the logger definition version for this group.}

\item{ResourceDefinitionVersionArn}{The resource definition version ARN for this group.}

\item{SubscriptionDefinitionVersionArn}{The ARN of the subscription definition version for this group.}
}
\description{
Creates a version of a group which has already been defined.
}
\examples{

}
