% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.greengrass_operations.R
\name{create_logger_definition_version}
\alias{create_logger_definition_version}
\title{Creates a version of a logger definition that has already been defined}
\usage{
create_logger_definition_version(AmznClientToken = NULL,
  LoggerDefinitionId, Loggers = NULL)
}
\arguments{
\item{AmznClientToken}{A client token used to correlate requests and responses.}

\item{LoggerDefinitionId}{[required] The ID of the logger definition.}

\item{Loggers}{A list of loggers.}
}
\description{
Creates a version of a logger definition that has already been defined.
}
\section{Accepted Parameters}{
\preformatted{create_logger_definition_version(
  AmznClientToken = "string",
  LoggerDefinitionId = "string",
  Loggers = list(
    list(
      Component = "GreengrassSystem"|"Lambda",
      Id = "string",
      Level = "DEBUG"|"INFO"|"WARN"|"ERROR"|"FATAL",
      Space = 123,
      Type = "FileSystem"|"AWSCloudWatch"
    )
  )
)
}
}

