% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.greengrass_operations.R
\name{create_resource_definition}
\alias{create_resource_definition}
\title{Creates a resource definition which contains a list of resources to be used in a group}
\usage{
create_resource_definition(AmznClientToken = NULL,
  InitialVersion = NULL, Name = NULL)
}
\arguments{
\item{AmznClientToken}{A client token used to correlate requests and responses.}

\item{InitialVersion}{Information about the initial version of the resource definition.}

\item{Name}{The name of the resource definition.}
}
\description{
Creates a resource definition which contains a list of resources to be used in a group. You can create an initial version of the definition by providing a list of resources now, or use ''CreateResourceDefinitionVersion'' later.
}
\section{Accepted Parameters}{
\preformatted{create_resource_definition(
  AmznClientToken = "string",
  InitialVersion = list(
    Resources = list(
      list(
        Id = "string",
        Name = "string",
        ResourceDataContainer = list(
          LocalDeviceResourceData = list(
            GroupOwnerSetting = list(
              AutoAddGroupOwner = TRUE|FALSE,
              GroupOwner = "string"
            ),
            SourcePath = "string"
          ),
          LocalVolumeResourceData = list(
            DestinationPath = "string",
            GroupOwnerSetting = list(
              AutoAddGroupOwner = TRUE|FALSE,
              GroupOwner = "string"
            ),
            SourcePath = "string"
          ),
          S3MachineLearningModelResourceData = list(
            DestinationPath = "string",
            S3Uri = "string"
          ),
          SageMakerMachineLearningModelResourceData = list(
            DestinationPath = "string",
            SageMakerJobArn = "string"
          ),
          SecretsManagerSecretResourceData = list(
            ARN = "string",
            AdditionalStagingLabelsToDownload = list(
              "string"
            )
          )
        )
      )
    )
  ),
  Name = "string"
)
}
}

