% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.greengrass_operations.R
\name{create_resource_definition_version}
\alias{create_resource_definition_version}
\title{Creates a version of a resource definition that has already been defined}
\usage{
create_resource_definition_version(ResourceDefinitionId,
  AmznClientToken = NULL, Resources = NULL)
}
\arguments{
\item{ResourceDefinitionId}{The ID of the resource definition.}

\item{AmznClientToken}{A client token used to correlate requests and responses.}

\item{Resources}{A list of resources.}
}
\description{
Creates a version of a resource definition that has already been defined.
}
\examples{

}
