% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.greengrass_operations.R
\name{create_subscription_definition}
\alias{create_subscription_definition}
\title{Creates a subscription definition}
\usage{
create_subscription_definition(AmznClientToken = NULL,
  InitialVersion = NULL, Name = NULL)
}
\arguments{
\item{AmznClientToken}{A client token used to correlate requests and responses.}

\item{InitialVersion}{Information about the initial version of the subscription definition.}

\item{Name}{The name of the subscription definition.}
}
\description{
Creates a subscription definition. You may provide the initial version of the subscription definition now or use ''CreateSubscriptionDefinitionVersion'' at a later time.
}
\examples{

}
