% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.greengrass_operations.R
\name{create_subscription_definition_version}
\alias{create_subscription_definition_version}
\title{Creates a version of a subscription definition which has already been defined}
\usage{
create_subscription_definition_version(SubscriptionDefinitionId,
  AmznClientToken = NULL, Subscriptions = NULL)
}
\arguments{
\item{SubscriptionDefinitionId}{The ID of the subscription definition.}

\item{AmznClientToken}{A client token used to correlate requests and responses.}

\item{Subscriptions}{A list of subscriptions.}
}
\description{
Creates a version of a subscription definition which has already been defined.
}
\examples{

}
