% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.greengrass_operations.R
\name{delete_connector_definition}
\alias{delete_connector_definition}
\title{Deletes a connector definition}
\usage{
delete_connector_definition(ConnectorDefinitionId)
}
\arguments{
\item{ConnectorDefinitionId}{[required] The ID of the connector definition.}
}
\description{
Deletes a connector definition.
}
\section{Accepted Parameters}{
\preformatted{delete_connector_definition(
  ConnectorDefinitionId = "string"
)
}
}

