% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.greengrass_operations.R
\name{delete_core_definition}
\alias{delete_core_definition}
\title{Deletes a core definition}
\usage{
delete_core_definition(
  CoreDefinitionId = "string"
)
}
\arguments{
\item{CoreDefinitionId}{[required] The ID of the core definition.}
}
\description{
Deletes a core definition.
}
\examples{

}
