% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.greengrass_operations.R
\name{delete_device_definition}
\alias{delete_device_definition}
\title{Deletes a device definition}
\usage{
delete_device_definition(DeviceDefinitionId)
}
\arguments{
\item{DeviceDefinitionId}{[required] The ID of the device definition.}
}
\description{
Deletes a device definition.
}
\section{Accepted Parameters}{
\preformatted{delete_device_definition(
  DeviceDefinitionId = "string"
)
}
}

