% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.greengrass_operations.R
\name{delete_function_definition}
\alias{delete_function_definition}
\title{Deletes a Lambda function definition}
\usage{
delete_function_definition(FunctionDefinitionId)
}
\arguments{
\item{FunctionDefinitionId}{The ID of the Lambda function definition.}
}
\description{
Deletes a Lambda function definition.
}
\examples{

}
