% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.greengrass_operations.R
\name{delete_group}
\alias{delete_group}
\title{Deletes a group}
\usage{
delete_group(
  GroupId = "string"
)
}
\arguments{
\item{GroupId}{[required] The ID of the Greengrass group.}
}
\description{
Deletes a group.
}
\examples{

}
