% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.greengrass_operations.R
\name{delete_logger_definition}
\alias{delete_logger_definition}
\title{Deletes a logger definition}
\usage{
delete_logger_definition(LoggerDefinitionId)
}
\arguments{
\item{LoggerDefinitionId}{The ID of the logger definition.}
}
\description{
Deletes a logger definition.
}
\examples{

}
