% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.greengrass_operations.R
\name{get_connectivity_info}
\alias{get_connectivity_info}
\title{Retrieves the connectivity information for a core}
\usage{
get_connectivity_info(
  ThingName = "string"
)
}
\arguments{
\item{ThingName}{[required] The thing name.}
}
\description{
Retrieves the connectivity information for a core.
}
\examples{

}
