% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.greengrass_operations.R
\name{get_connector_definition}
\alias{get_connector_definition}
\title{Retrieves information about a connector definition}
\usage{
get_connector_definition(
  ConnectorDefinitionId = "string"
)
}
\arguments{
\item{ConnectorDefinitionId}{[required] The ID of the connector definition.}
}
\description{
Retrieves information about a connector definition.
}
\examples{

}
