% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.greengrass_operations.R
\name{get_connector_definition_version}
\alias{get_connector_definition_version}
\title{Retrieves information about a connector definition version, including the connectors that the version contains}
\usage{
get_connector_definition_version(
  ConnectorDefinitionId = "string",
  ConnectorDefinitionVersionId = "string",
  NextToken = "string"
)
}
\arguments{
\item{ConnectorDefinitionId}{[required] The ID of the connector definition.}

\item{ConnectorDefinitionVersionId}{[required] The ID of the connector definition version.}

\item{NextToken}{The token for the next set of results, or ''null'' if there are no additional results.}
}
\description{
Retrieves information about a connector definition version, including the connectors that the version contains. Connectors are prebuilt modules that interact with local infrastructure, device protocols, AWS, and other cloud services.
}
\examples{

}
