% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.greengrass_operations.R
\name{get_core_definition}
\alias{get_core_definition}
\title{Retrieves information about a core definition version}
\usage{
get_core_definition(CoreDefinitionId)
}
\arguments{
\item{CoreDefinitionId}{The ID of the core definition.}
}
\description{
Retrieves information about a core definition version.
}
\examples{

}
