% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.greengrass_operations.R
\name{get_core_definition_version}
\alias{get_core_definition_version}
\title{Retrieves information about a core definition version}
\usage{
get_core_definition_version(CoreDefinitionId, CoreDefinitionVersionId)
}
\arguments{
\item{CoreDefinitionId}{The ID of the core definition.}

\item{CoreDefinitionVersionId}{The ID of the core definition version.}
}
\description{
Retrieves information about a core definition version.
}
\examples{

}
