% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.greengrass_operations.R
\name{get_deployment_status}
\alias{get_deployment_status}
\title{Returns the status of a deployment}
\usage{
get_deployment_status(GroupId, DeploymentId)
}
\arguments{
\item{GroupId}{The ID of the AWS Greengrass group.}

\item{DeploymentId}{The ID of the deployment.}
}
\description{
Returns the status of a deployment.
}
\examples{

}
