% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.greengrass_operations.R
\name{get_device_definition}
\alias{get_device_definition}
\title{Retrieves information about a device definition}
\usage{
get_device_definition(DeviceDefinitionId)
}
\arguments{
\item{DeviceDefinitionId}{[required] The ID of the device definition.}
}
\description{
Retrieves information about a device definition.
}
\section{Accepted Parameters}{
\preformatted{get_device_definition(
  DeviceDefinitionId = "string"
)
}
}

