% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.greengrass_operations.R
\name{get_device_definition_version}
\alias{get_device_definition_version}
\title{Retrieves information about a device definition version}
\usage{
get_device_definition_version(DeviceDefinitionId,
  DeviceDefinitionVersionId, NextToken = NULL)
}
\arguments{
\item{DeviceDefinitionId}{[required] The ID of the device definition.}

\item{DeviceDefinitionVersionId}{[required] The ID of the device definition version.}

\item{NextToken}{The token for the next set of results, or ''null'' if there are no additional results.}
}
\description{
Retrieves information about a device definition version.
}
\section{Accepted Parameters}{
\preformatted{get_device_definition_version(
  DeviceDefinitionId = "string",
  DeviceDefinitionVersionId = "string",
  NextToken = "string"
)
}
}

