% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.greengrass_operations.R
\name{get_function_definition}
\alias{get_function_definition}
\title{Retrieves information about a Lambda function definition, including its creation time and latest version}
\usage{
get_function_definition(FunctionDefinitionId)
}
\arguments{
\item{FunctionDefinitionId}{The ID of the Lambda function definition.}
}
\description{
Retrieves information about a Lambda function definition, including its creation time and latest version.
}
\examples{

}
