% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.greengrass_operations.R
\name{get_function_definition_version}
\alias{get_function_definition_version}
\title{Retrieves information about a Lambda function definition version, including which Lambda functions are included in the version and their configurations}
\usage{
get_function_definition_version(FunctionDefinitionId,
  FunctionDefinitionVersionId)
}
\arguments{
\item{FunctionDefinitionId}{The ID of the Lambda function definition.}

\item{FunctionDefinitionVersionId}{The ID of the function definition version.}
}
\description{
Retrieves information about a Lambda function definition version, including which Lambda functions are included in the version and their configurations.
}
\examples{

}
