% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.greengrass_operations.R
\name{get_group}
\alias{get_group}
\title{Retrieves information about a group}
\usage{
get_group(GroupId)
}
\arguments{
\item{GroupId}{[required] The ID of the AWS Greengrass group.}
}
\description{
Retrieves information about a group.
}
\section{Accepted Parameters}{
\preformatted{get_group(
  GroupId = "string"
)
}
}

