% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.greengrass_operations.R
\name{get_group_certificate_authority}
\alias{get_group_certificate_authority}
\title{Retreives the CA associated with a group}
\usage{
get_group_certificate_authority(CertificateAuthorityId, GroupId)
}
\arguments{
\item{CertificateAuthorityId}{The ID of the certificate authority.}

\item{GroupId}{The ID of the AWS Greengrass group.}
}
\description{
Retreives the CA associated with a group. Returns the public key of the CA.
}
\examples{

}
