% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.greengrass_operations.R
\name{get_group_certificate_configuration}
\alias{get_group_certificate_configuration}
\title{Retrieves the current configuration for the CA used by the group}
\usage{
get_group_certificate_configuration(GroupId)
}
\arguments{
\item{GroupId}{[required] The ID of the AWS Greengrass group.}
}
\description{
Retrieves the current configuration for the CA used by the group.
}
\section{Accepted Parameters}{
\preformatted{get_group_certificate_configuration(
  GroupId = "string"
)
}
}

