% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.greengrass_operations.R
\name{get_group_version}
\alias{get_group_version}
\title{Retrieves information about a group version}
\usage{
get_group_version(GroupVersionId, GroupId)
}
\arguments{
\item{GroupVersionId}{The ID of the group version.}

\item{GroupId}{The ID of the AWS Greengrass group.}
}
\description{
Retrieves information about a group version.
}
\examples{

}
