% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.greengrass_operations.R
\name{get_logger_definition}
\alias{get_logger_definition}
\title{Retrieves information about a logger definition}
\usage{
get_logger_definition(
  LoggerDefinitionId = "string"
)
}
\arguments{
\item{LoggerDefinitionId}{[required] The ID of the logger definition.}
}
\description{
Retrieves information about a logger definition.
}
\examples{

}
