% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.greengrass_operations.R
\name{get_logger_definition_version}
\alias{get_logger_definition_version}
\title{Retrieves information about a logger definition version}
\usage{
get_logger_definition_version(LoggerDefinitionVersionId, LoggerDefinitionId)
}
\arguments{
\item{LoggerDefinitionVersionId}{The ID of the logger definition version.}

\item{LoggerDefinitionId}{The ID of the logger definition.}
}
\description{
Retrieves information about a logger definition version.
}
\examples{

}
