% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.greengrass_operations.R
\name{get_resource_definition_version}
\alias{get_resource_definition_version}
\title{Retrieves information about a resource definition version, including which resources are included in the version}
\usage{
get_resource_definition_version(ResourceDefinitionVersionId,
  ResourceDefinitionId)
}
\arguments{
\item{ResourceDefinitionVersionId}{The ID of the resource definition version.}

\item{ResourceDefinitionId}{The ID of the resource definition.}
}
\description{
Retrieves information about a resource definition version, including which resources are included in the version.
}
\examples{

}
