% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.greengrass_operations.R
\name{get_service_role_for_account}
\alias{get_service_role_for_account}
\title{Retrieves the service role that is attached to your account}
\usage{
get_service_role_for_account()
}
\description{
Retrieves the service role that is attached to your account.
}
\section{Accepted Parameters}{
\preformatted{get_service_role_for_account()
}
}

