% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.greengrass_operations.R
\name{get_subscription_definition_version}
\alias{get_subscription_definition_version}
\title{Retrieves information about a subscription definition version}
\usage{
get_subscription_definition_version(SubscriptionDefinitionId,
  SubscriptionDefinitionVersionId)
}
\arguments{
\item{SubscriptionDefinitionId}{The ID of the subscription definition.}

\item{SubscriptionDefinitionVersionId}{The ID of the subscription definition version.}
}
\description{
Retrieves information about a subscription definition version.
}
\examples{

}
