% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.greengrass_operations.R
\name{list_bulk_deployment_detailed_reports}
\alias{list_bulk_deployment_detailed_reports}
\title{Gets a paginated list of the deployments that have been started in a bulk deployment operation, and their current deployment status}
\usage{
list_bulk_deployment_detailed_reports(
  BulkDeploymentId = "string",
  MaxResults = "string",
  NextToken = "string"
)
}
\arguments{
\item{BulkDeploymentId}{[required] The ID of the bulk deployment.}

\item{MaxResults}{The maximum number of results to be returned per request.}

\item{NextToken}{The token for the next set of results, or ''null'' if there are no additional results.}
}
\description{
Gets a paginated list of the deployments that have been started in a bulk deployment operation, and their current deployment status.
}
\examples{

}
