% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.greengrass_operations.R
\name{list_bulk_deployments}
\alias{list_bulk_deployments}
\title{Returns a list of bulk deployments}
\usage{
list_bulk_deployments(MaxResults = NULL, NextToken = NULL)
}
\arguments{
\item{MaxResults}{The maximum number of results to be returned per request.}

\item{NextToken}{The token for the next set of results, or ''null'' if there are no additional results.}
}
\description{
Returns a list of bulk deployments.
}
\section{Accepted Parameters}{
\preformatted{list_bulk_deployments(
  MaxResults = "string",
  NextToken = "string"
)
}
}

