% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.greengrass_operations.R
\name{list_connector_definition_versions}
\alias{list_connector_definition_versions}
\title{Lists the versions of a connector definition, which are containers for connectors}
\usage{
list_connector_definition_versions(
  ConnectorDefinitionId = "string",
  MaxResults = "string",
  NextToken = "string"
)
}
\arguments{
\item{ConnectorDefinitionId}{[required] The ID of the connector definition.}

\item{MaxResults}{The maximum number of results to be returned per request.}

\item{NextToken}{The token for the next set of results, or ''null'' if there are no additional results.}
}
\description{
Lists the versions of a connector definition, which are containers for connectors. Connectors run on the Greengrass core and contain built-in integration with local infrastructure, device protocols, AWS, and other cloud services.
}
\examples{

}
