% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.greengrass_operations.R
\name{list_deployments}
\alias{list_deployments}
\title{Returns a history of deployments for the group}
\usage{
list_deployments(
  GroupId = "string",
  MaxResults = "string",
  NextToken = "string"
)
}
\arguments{
\item{GroupId}{[required] The ID of the Greengrass group.}

\item{MaxResults}{The maximum number of results to be returned per request.}

\item{NextToken}{The token for the next set of results, or ''null'' if there are no additional results.}
}
\description{
Returns a history of deployments for the group.
}
\examples{

}
