% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.greengrass_operations.R
\name{list_device_definitions}
\alias{list_device_definitions}
\title{Retrieves a list of device definitions}
\usage{
list_device_definitions(
  MaxResults = "string",
  NextToken = "string"
)
}
\arguments{
\item{MaxResults}{The maximum number of results to be returned per request.}

\item{NextToken}{The token for the next set of results, or ''null'' if there are no additional results.}
}
\description{
Retrieves a list of device definitions.
}
\examples{

}
