% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.greengrass_operations.R
\name{list_group_certificate_authorities}
\alias{list_group_certificate_authorities}
\title{Retrieves the current CAs for a group}
\usage{
list_group_certificate_authorities(GroupId)
}
\arguments{
\item{GroupId}{[required] The ID of the AWS Greengrass group.}
}
\description{
Retrieves the current CAs for a group.
}
\section{Accepted Parameters}{
\preformatted{list_group_certificate_authorities(
  GroupId = "string"
)
}
}

