% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.greengrass_operations.R
\name{list_group_versions}
\alias{list_group_versions}
\title{Lists the versions of a group}
\usage{
list_group_versions(GroupId, MaxResults = NULL, NextToken = NULL)
}
\arguments{
\item{GroupId}{[required] The ID of the Greengrass group.}

\item{MaxResults}{The maximum number of results to be returned per request.}

\item{NextToken}{The token for the next set of results, or ''null'' if there are no additional results.}
}
\description{
Lists the versions of a group.
}
\section{Accepted Parameters}{
\preformatted{list_group_versions(
  GroupId = "string",
  MaxResults = "string",
  NextToken = "string"
)
}
}

