% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.greengrass_operations.R
\name{list_logger_definition_versions}
\alias{list_logger_definition_versions}
\title{Lists the versions of a logger definition}
\usage{
list_logger_definition_versions(LoggerDefinitionId, MaxResults = NULL,
  NextToken = NULL)
}
\arguments{
\item{LoggerDefinitionId}{The ID of the logger definition.}

\item{MaxResults}{The maximum number of results to be returned per request.}

\item{NextToken}{The token for the next set of results, or ''null'' if there are no additional results.}
}
\description{
Lists the versions of a logger definition.
}
\examples{

}
