% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.greengrass_operations.R
\name{list_resource_definitions}
\alias{list_resource_definitions}
\title{Retrieves a list of resource definitions}
\usage{
list_resource_definitions(MaxResults = NULL, NextToken = NULL)
}
\arguments{
\item{MaxResults}{The maximum number of results to be returned per request.}

\item{NextToken}{The token for the next set of results, or ''null'' if there are no additional results.}
}
\description{
Retrieves a list of resource definitions.
}
\section{Accepted Parameters}{
\preformatted{list_resource_definitions(
  MaxResults = "string",
  NextToken = "string"
)
}
}

