% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.greengrass_package.R
\docType{package}
\name{paws.greengrass-package}
\alias{paws.greengrass}
\alias{paws.greengrass-package}
\title{paws.greengrass: AWS Greengrass}
\description{
AWS IoT Greengrass seamlessly extends AWS onto
    physical devices so they can act locally on the data they generate,
    while still using the cloud for management, analytics, and durable
    storage. AWS IoT Greengrass ensures your devices can respond quickly
    to local events and operate with intermittent connectivity. AWS IoT
    Greengrass minimizes the cost of transmitting data to the cloud by
    allowing you to author AWS Lambda functions that execute locally.
}
\author{
\strong{Maintainer}: David Kretch \email{david.kretch@gmail.com}

Authors:
\itemize{
  \item Adam Banker \email{adam.banker39@gmail.com}
}

Other contributors:
\itemize{
  \item Amazon.com, Inc. [copyright holder]
}

}
\keyword{internal}
