% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.greengrass_operations.R
\name{update_connector_definition}
\alias{update_connector_definition}
\title{Updates a connector definition}
\usage{
update_connector_definition(ConnectorDefinitionId, Name = NULL)
}
\arguments{
\item{ConnectorDefinitionId}{[required] The ID of the connector definition.}

\item{Name}{The name of the definition.}
}
\description{
Updates a connector definition.
}
\section{Accepted Parameters}{
\preformatted{update_connector_definition(
  ConnectorDefinitionId = "string",
  Name = "string"
)
}
}

