% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.greengrass_operations.R
\name{update_core_definition}
\alias{update_core_definition}
\title{Updates a core definition}
\usage{
update_core_definition(
  CoreDefinitionId = "string",
  Name = "string"
)
}
\arguments{
\item{CoreDefinitionId}{[required] The ID of the core definition.}

\item{Name}{The name of the definition.}
}
\description{
Updates a core definition.
}
\examples{

}
