% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.greengrass_operations.R
\name{update_device_definition}
\alias{update_device_definition}
\title{Updates a device definition}
\usage{
update_device_definition(DeviceDefinitionId, Name = NULL)
}
\arguments{
\item{DeviceDefinitionId}{[required] The ID of the device definition.}

\item{Name}{The name of the definition.}
}
\description{
Updates a device definition.
}
\section{Accepted Parameters}{
\preformatted{update_device_definition(
  DeviceDefinitionId = "string",
  Name = "string"
)
}
}

