% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.greengrass_operations.R
\name{update_function_definition}
\alias{update_function_definition}
\title{Updates a Lambda function definition}
\usage{
update_function_definition(FunctionDefinitionId, Name = NULL)
}
\arguments{
\item{FunctionDefinitionId}{[required] The ID of the Lambda function definition.}

\item{Name}{The name of the definition.}
}
\description{
Updates a Lambda function definition.
}
\section{Accepted Parameters}{
\preformatted{update_function_definition(
  FunctionDefinitionId = "string",
  Name = "string"
)
}
}

