% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.greengrass_operations.R
\name{update_group_certificate_configuration}
\alias{update_group_certificate_configuration}
\title{Updates the Certificate expiry time for a group}
\usage{

  update_group_certificate_configuration(CertificateExpiryInMilliseconds = NULL,
  GroupId)
}
\arguments{
\item{CertificateExpiryInMilliseconds}{The amount of time remaining before the certificate expires, in milliseconds.}

\item{GroupId}{[required] The ID of the AWS Greengrass group.}
}
\description{
Updates the Certificate expiry time for a group.
}
\section{Accepted Parameters}{
\preformatted{update_group_certificate_configuration(
  CertificateExpiryInMilliseconds = "string",
  GroupId = "string"
)
}
}

