% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.greengrass_operations.R
\name{update_logger_definition}
\alias{update_logger_definition}
\title{Updates a logger definition}
\usage{
update_logger_definition(LoggerDefinitionId, Name = NULL)
}
\arguments{
\item{LoggerDefinitionId}{[required] The ID of the logger definition.}

\item{Name}{The name of the definition.}
}
\description{
Updates a logger definition.
}
\section{Accepted Parameters}{
\preformatted{update_logger_definition(
  LoggerDefinitionId = "string",
  Name = "string"
)
}
}

