% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.greengrass_operations.R
\name{update_resource_definition}
\alias{update_resource_definition}
\title{Updates a resource definition}
\usage{
update_resource_definition(Name = NULL, ResourceDefinitionId)
}
\arguments{
\item{Name}{The name of the definition.}

\item{ResourceDefinitionId}{[required] The ID of the resource definition.}
}
\description{
Updates a resource definition.
}
\section{Accepted Parameters}{
\preformatted{update_resource_definition(
  Name = "string",
  ResourceDefinitionId = "string"
)
}
}

