% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.greengrass_operations.R
\name{update_subscription_definition}
\alias{update_subscription_definition}
\title{Updates a subscription definition}
\usage{
update_subscription_definition(SubscriptionDefinitionId, Name = NULL)
}
\arguments{
\item{SubscriptionDefinitionId}{The ID of the subscription definition.}

\item{Name}{The name of the definition.}
}
\description{
Updates a subscription definition.
}
\examples{

}
