% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.greengrass_operations.R
\name{create_function_definition_version}
\alias{create_function_definition_version}
\title{Creates a version of a Lambda function definition that has already been defined}
\usage{
create_function_definition_version(FunctionDefinitionId,
  AmznClientToken = NULL, Functions = NULL)
}
\arguments{
\item{FunctionDefinitionId}{The ID of the Lambda function definition.}

\item{AmznClientToken}{A client token used to correlate requests and responses.}

\item{Functions}{A list of Lambda functions in this function definition version.}
}
\description{
Creates a version of a Lambda function definition that has already been defined.
}
\examples{

}
