% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.greengrass_operations.R
\name{create_group_certificate_authority}
\alias{create_group_certificate_authority}
\title{Creates a CA for the group}
\usage{
create_group_certificate_authority(AmznClientToken = NULL, GroupId)
}
\arguments{
\item{AmznClientToken}{A client token used to correlate requests and responses.}

\item{GroupId}{[required] The ID of the Greengrass group.}
}
\description{
Creates a CA for the group. If a CA already exists, it will rotate the existing CA.
}
\section{Accepted Parameters}{
\preformatted{create_group_certificate_authority(
  AmznClientToken = "string",
  GroupId = "string"
)
}
}

