% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.greengrass_operations.R
\name{create_logger_definition_version}
\alias{create_logger_definition_version}
\title{Creates a version of a logger definition that has already been defined}
\usage{
create_logger_definition_version(LoggerDefinitionId, AmznClientToken = NULL,
  Loggers = NULL)
}
\arguments{
\item{LoggerDefinitionId}{The ID of the logger definition.}

\item{AmznClientToken}{A client token used to correlate requests and responses.}

\item{Loggers}{A list of loggers.}
}
\description{
Creates a version of a logger definition that has already been defined.
}
\examples{

}
