% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.greengrass_operations.R
\name{delete_resource_definition}
\alias{delete_resource_definition}
\title{Deletes a resource definition}
\usage{
delete_resource_definition(ResourceDefinitionId)
}
\arguments{
\item{ResourceDefinitionId}{The ID of the resource definition.}
}
\description{
Deletes a resource definition.
}
\examples{

}
