% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.greengrass_operations.R
\name{delete_subscription_definition}
\alias{delete_subscription_definition}
\title{Deletes a subscription definition}
\usage{
delete_subscription_definition(SubscriptionDefinitionId)
}
\arguments{
\item{SubscriptionDefinitionId}{[required] The ID of the subscription definition.}
}
\description{
Deletes a subscription definition.
}
\section{Accepted Parameters}{
\preformatted{delete_subscription_definition(
  SubscriptionDefinitionId = "string"
)
}
}

