% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.greengrass_operations.R
\name{get_associated_role}
\alias{get_associated_role}
\title{Retrieves the role associated with a particular group}
\usage{
get_associated_role(
  GroupId = "string"
)
}
\arguments{
\item{GroupId}{[required] The ID of the Greengrass group.}
}
\description{
Retrieves the role associated with a particular group.
}
\examples{

}
