% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.greengrass_operations.R
\name{get_bulk_deployment_status}
\alias{get_bulk_deployment_status}
\title{Returns the status of a bulk deployment}
\usage{
get_bulk_deployment_status(BulkDeploymentId)
}
\arguments{
\item{BulkDeploymentId}{[required] The ID of the bulk deployment.}
}
\description{
Returns the status of a bulk deployment.
}
\section{Accepted Parameters}{
\preformatted{get_bulk_deployment_status(
  BulkDeploymentId = "string"
)
}
}

