% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.greengrass_operations.R
\name{get_deployment_status}
\alias{get_deployment_status}
\title{Returns the status of a deployment}
\usage{
get_deployment_status(DeploymentId, GroupId)
}
\arguments{
\item{DeploymentId}{[required] The ID of the deployment.}

\item{GroupId}{[required] The ID of the Greengrass group.}
}
\description{
Returns the status of a deployment.
}
\section{Accepted Parameters}{
\preformatted{get_deployment_status(
  DeploymentId = "string",
  GroupId = "string"
)
}
}

