% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.greengrass_operations.R
\name{get_device_definition_version}
\alias{get_device_definition_version}
\title{Retrieves information about a device definition version}
\usage{
get_device_definition_version(DeviceDefinitionVersionId, DeviceDefinitionId)
}
\arguments{
\item{DeviceDefinitionVersionId}{The ID of the device definition version.}

\item{DeviceDefinitionId}{The ID of the device definition.}
}
\description{
Retrieves information about a device definition version.
}
\examples{

}
