% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.greengrass_operations.R
\name{get_function_definition_version}
\alias{get_function_definition_version}
\title{Retrieves information about a Lambda function definition version, including which Lambda functions are included in the version and their configurations}
\usage{
get_function_definition_version(FunctionDefinitionId,
  FunctionDefinitionVersionId, NextToken = NULL)
}
\arguments{
\item{FunctionDefinitionId}{[required] The ID of the Lambda function definition.}

\item{FunctionDefinitionVersionId}{[required] The ID of the function definition version.}

\item{NextToken}{The token for the next set of results, or ''null'' if there are no additional results.}
}
\description{
Retrieves information about a Lambda function definition version, including which Lambda functions are included in the version and their configurations.
}
\section{Accepted Parameters}{
\preformatted{get_function_definition_version(
  FunctionDefinitionId = "string",
  FunctionDefinitionVersionId = "string",
  NextToken = "string"
)
}
}

