% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.greengrass_operations.R
\name{get_group_certificate_authority}
\alias{get_group_certificate_authority}
\title{Retreives the CA associated with a group}
\usage{
get_group_certificate_authority(
  CertificateAuthorityId = "string",
  GroupId = "string"
)
}
\arguments{
\item{CertificateAuthorityId}{[required] The ID of the certificate authority.}

\item{GroupId}{[required] The ID of the Greengrass group.}
}
\description{
Retreives the CA associated with a group. Returns the public key of the CA.
}
\examples{

}
