% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.greengrass_operations.R
\name{get_group_version}
\alias{get_group_version}
\title{Retrieves information about a group version}
\usage{
get_group_version(GroupId, GroupVersionId)
}
\arguments{
\item{GroupId}{[required] The ID of the Greengrass group.}

\item{GroupVersionId}{[required] The ID of the group version.}
}
\description{
Retrieves information about a group version.
}
\section{Accepted Parameters}{
\preformatted{get_group_version(
  GroupId = "string",
  GroupVersionId = "string"
)
}
}

