% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.greengrass_operations.R
\name{get_resource_definition}
\alias{get_resource_definition}
\title{Retrieves information about a resource definition, including its creation time and latest version}
\usage{
get_resource_definition(
  ResourceDefinitionId = "string"
)
}
\arguments{
\item{ResourceDefinitionId}{[required] The ID of the resource definition.}
}
\description{
Retrieves information about a resource definition, including its creation time and latest version.
}
\examples{

}
