% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.greengrass_operations.R
\name{get_subscription_definition}
\alias{get_subscription_definition}
\title{Retrieves information about a subscription definition}
\usage{
get_subscription_definition(SubscriptionDefinitionId)
}
\arguments{
\item{SubscriptionDefinitionId}{The ID of the subscription definition.}
}
\description{
Retrieves information about a subscription definition.
}
\examples{

}
