% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.greengrass_operations.R
\name{get_subscription_definition_version}
\alias{get_subscription_definition_version}
\title{Retrieves information about a subscription definition version}
\usage{
get_subscription_definition_version(
  NextToken = "string",
  SubscriptionDefinitionId = "string",
  SubscriptionDefinitionVersionId = "string"
)
}
\arguments{
\item{NextToken}{The token for the next set of results, or ''null'' if there are no additional results.}

\item{SubscriptionDefinitionId}{[required] The ID of the subscription definition.}

\item{SubscriptionDefinitionVersionId}{[required] The ID of the subscription definition version.}
}
\description{
Retrieves information about a subscription definition version.
}
\examples{

}
