% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.greengrass_operations.R
\name{list_resource_definition_versions}
\alias{list_resource_definition_versions}
\title{Lists the versions of a resource definition}
\usage{
list_resource_definition_versions(ResourceDefinitionId, MaxResults = NULL,
  NextToken = NULL)
}
\arguments{
\item{ResourceDefinitionId}{The ID of the resource definition.}

\item{MaxResults}{The maximum number of results to be returned per request.}

\item{NextToken}{The token for the next set of results, or ''null'' if there are no additional results.}
}
\description{
Lists the versions of a resource definition.
}
\examples{

}
