% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.greengrass_operations.R
\name{list_subscription_definition_versions}
\alias{list_subscription_definition_versions}
\title{Lists the versions of a subscription definition}
\usage{
list_subscription_definition_versions(SubscriptionDefinitionId,
  MaxResults = NULL, NextToken = NULL)
}
\arguments{
\item{SubscriptionDefinitionId}{The ID of the subscription definition.}

\item{MaxResults}{The maximum number of results to be returned per request.}

\item{NextToken}{The token for the next set of results, or ''null'' if there are no additional results.}
}
\description{
Lists the versions of a subscription definition.
}
\examples{

}
