% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.greengrass_operations.R
\name{update_connectivity_info}
\alias{update_connectivity_info}
\title{Updates the connectivity information for the core}
\usage{
update_connectivity_info(ThingName, ConnectivityInfo = NULL)
}
\arguments{
\item{ThingName}{The thing name.}

\item{ConnectivityInfo}{A list of connectivity info.}
}
\description{
Updates the connectivity information for the core. Any devices that belong to the group which has this core will receive this information in order to find the location of the core and connect to it.
}
\examples{

}
