% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.greengrass_operations.R
\name{update_group}
\alias{update_group}
\title{Updates a group}
\usage{
update_group(GroupId, Name = NULL)
}
\arguments{
\item{GroupId}{The ID of the AWS Greengrass group.}

\item{Name}{The name of the definition.}
}
\description{
Updates a group.
}
\examples{

}
