% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.greengrass_operations.R
\name{update_group}
\alias{update_group}
\title{Updates a group}
\usage{
update_group(
  GroupId = "string",
  Name = "string"
)
}
\arguments{
\item{GroupId}{[required] The ID of the Greengrass group.}

\item{Name}{The name of the definition.}
}
\description{
Updates a group.
}
\examples{

}
