% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.greengrass_operations.R
\name{update_group_certificate_configuration}
\alias{update_group_certificate_configuration}
\title{Updates the Certificate expiry time for a group}
\usage{
update_group_certificate_configuration(GroupId,
  CertificateExpiryInMilliseconds = NULL)
}
\arguments{
\item{GroupId}{The ID of the AWS Greengrass group.}

\item{CertificateExpiryInMilliseconds}{The amount of time remaining before the certificate expires, in milliseconds.}
}
\description{
Updates the Certificate expiry time for a group.
}
\examples{

}
