% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.greengrass_operations.R
\name{update_resource_definition}
\alias{update_resource_definition}
\title{Updates a resource definition}
\usage{
update_resource_definition(ResourceDefinitionId, Name = NULL)
}
\arguments{
\item{ResourceDefinitionId}{The ID of the resource definition.}

\item{Name}{The name of the definition.}
}
\description{
Updates a resource definition.
}
\examples{

}
