% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.greengrass_operations.R
\name{update_subscription_definition}
\alias{update_subscription_definition}
\title{Updates a subscription definition}
\usage{
update_subscription_definition(
  Name = "string",
  SubscriptionDefinitionId = "string"
)
}
\arguments{
\item{Name}{The name of the definition.}

\item{SubscriptionDefinitionId}{[required] The ID of the subscription definition.}
}
\description{
Updates a subscription definition.
}
\examples{

}
